rename interface from signer to keys
